Move Auth0 things to its own module
