+++ app/src/main/java/com/lambdaschool/hackathon_portal/di/app/AppComponent.kt
@@ -3,6 +3,7 @@ package com.lambdaschool.hackathon_portal.di.app
 import android.app.Application
 import com.lambdaschool.hackathon_portal.di.app.modules.AppModule
 import com.lambdaschool.hackathon_portal.di.activity.ActivityComponent
+import com.lambdaschool.hackathon_portal.di.app.modules.Auth0Module
 import com.lambdaschool.hackathon_portal.di.app.modules.ViewModelsModule
 import dagger.BindsInstance
 import dagger.Component
@@ -12,6 +13,7 @@ import javax.inject.Singleton
 @Component(
     modules = [
         AppModule::class,
+        Auth0Module::class,
         ViewModelsModule::class
     ])
 interface AppComponent {
