+++ app/src/main/java/com/lambdaschool/hackathon_portal/di/app/modules/AppModule.kt
@@ -1,10 +1,5 @@
 package com.lambdaschool.hackathon_portal.di.app.modules
 
-import android.app.Application
-import com.auth0.android.Auth0
-import com.auth0.android.authentication.AuthenticationAPIClient
-import com.auth0.android.authentication.storage.SecureCredentialsManager
-import com.auth0.android.authentication.storage.SharedPreferencesStorage
 import com.lambdaschool.hackathon_portal.repository.HackathonRepository
 import com.lambdaschool.hackathon_portal.retrofit.HackathonApiInterface
 import dagger.Module
@@ -33,24 +28,6 @@ object AppModule {
     @JvmStatic
     fun provideHackathonService(retrofit: Retrofit): HackathonApiInterface =
         retrofit.create(HackathonApiInterface::class.java)
-
-    @Singleton
-    @Provides
-    @JvmStatic
-    fun provideAuth0(application: Application): Auth0 {
-        val auth0 = Auth0(application.applicationContext)
-        auth0.isOIDCConformant = true
-        return auth0
-    }
-
-    @Singleton
-    @Provides
-    @JvmStatic
-    fun provideSecureCredentialsManager(application: Application, auth0: Auth0) =
-        SecureCredentialsManager(
-            application.applicationContext,
-            AuthenticationAPIClient(auth0),
-            SharedPreferencesStorage(application.applicationContext))
 
     @Singleton
     @Provides
